Fixed bug caused by not passing scope in wrapNodeFire() (#2)
The example in the README actually doesn't work out of the box because the scope is not passed along when `root()` is called in `theUser: stuff.root().child('users/{baz.qux}').get()`. The fix is to pass along `this.$scope` within `wrapNodeFire()` so each method has access to that same scope.