Properly set up lint, delint, and improve type usage.
Getting the type rights revealed a potential issue:  NodeFire can wrap
either a Reference or a Query, and didn't properly take into account
what would happen if it tried to delegate reference methods to a wrapped
query.  Fixed by always doing this.$ref.ref, which is a no-op for a
reference, but will return the underlying reference from a query.