Redefine Value type as NonNullable<unknown>|null
The previous type defintion involved an index signature, and therefore
any compatible object type was required to have an index signature which
made it incompatible with most object types. The idea was to guard
against objects that have a property with the value of undefined
somewhere down their object tree. But as it turns out this isn't
possible without causing false negatives, this was a mistake.