Deprecate type Value and alias it as 'any' for backwards compatibility
If Value is an object it is required to have an index signature which
makes it incompatible with most object types. The idea was to guard
against objects that have a property with the value of undefined
somewhere down their object tree. But as it turns out this isn't
possible without causing false negatives, this was a mistake.